refactor(authoring): merge the two RiggingScalarRows, keeping the material semantics
R1 of the editor refactoring plan. 1,446 lines across two files become 1,200
across three: one 344-line `inspector/RiggingScalarRow.tsx` plus 302 lines
deleted from each section. Net -246.

The plan assumed this was "the same component twice" and that the merge could be
behaviour-preserving. Diffing the bodies says otherwise — they had diverged in
three ways, and two of them change behaviour. The material copy is kept because on
every point it was the more considered implementation, so the **morph section
gains**:

1. Edits clamp to the constraint range again. The morph copy computed
   `minVal`/`maxVal` and then passed neither to `CommitOnBlurNumberInput`, so
   nothing bounded a typed current or default value.
2. A blocked binding is treated as unbound. The morph copy keyed min/max and
   `hasDifferentDefault` off "has input metadata" instead of off `isBound`, so a
   binding carrying a `blockedReason` still read its range from the standard input
   and could still show the modified-from-default dot.

The third difference was inert and stays that way: the morph copy defaulted an
absent constraint to `0` rather than `undefined`, but the only consumer renders
`typeof val === "number" ? val : 0`, so both showed `0`. It mattered solely
through the clamping the morph copy did not do.

`feature` is typed `SceneObjectFeature` again; the morph copy had degraded to
`any`.

Test coverage moves the right way. `RiggingMaterialSection` has no test file at
all, so the more careful implementation was the untested one; the two
`RiggingMorphTargetsSection` tests (locking, commit-on-blur) now exercise it.
741 passing, typecheck and lint clean.

`RiggingScalarRowProps` had to be exported: `RiggingColorRow` in the material
section takes the same prop set. That name is now inherited rather than chosen,
and is flagged in the docblock for whenever the colour and vector rows get the
same treatment.

The merged component stays in the feature layer rather than moving to `editor/` —
it calls `useInspectorTargetLock`, so it is store-bound by the plan's own rule.
Its reusable parts were already extracted: `RiggingPropertyRow` for the chassis,
`ChannelLockButton` for the lock, `useRowLock` for the aggregation.
